requirements.txt
Django==3.2
django-bootstrap3==10.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
coreapi
coreschema==0.0.4
psycopg2-binary
channels==3.0.4
drf-spectacular